updating version 04 Comments from Tero
z) I  

a) DH

<c>22</c><c>1024-bit MODP Group with 160-bit Prime Order Subgroup</c><c>SHOULD NOT</c>
<c>23</c><c>1024-bit MODP Group with 224-bit Prime Order Subgroup</c><c>SHOULD NOT</c>
<c>24</c><c>1024-bit MODP Group with 256-bit Prime Order Subgroup</c><c>SHOULD NOT</c>

<t> Group 5 or 1536-bit MODP Group is downgrade from MAY to SHOULD NOT. It was specified earlier,
but now considered to be vulnerable to be broken within the next few years by a nation state level attack,
so its security margin is considered too narrow.</t>

<t>Group 22-24 or 1024-bit MODP Group with 160-bit and 2048-bit MODP Group with 224-256-bit Prime Order 
Subgroup are envisioned to become small subgroups in a near future. It is expected in the
near future to be further downgraded to MUST NOT.</t>

COMMENT: Should we add any reference to the checks that needs to be performed.

COMMENT: Should we add in the DH intro that DH needs to implement strong security as to preserve perfect forward secrecy.

b) auth

<c>AUTH_HMAC_SHA1_96</c><c>MUST-</c><c></c>
<t> AUTH_HMAC_SHA1_96 has been downgraded from MUST in RFC4307 to MUST-. 
        There is an industry-wide trend to deprecate its usage.</t>

COMMENT: This leaves a MUST- notation. The other alternative could be to have SHOULD with the additional text specifying the next step is expected to be SHOULD NOT. 

c) prf

<c>PRF_HMAC_MD5</c><c>MUST NOT</c><c></c>
COMMENT: should we have a reference to https://www.mitls.org/downloads/transcript-collisions.pdf 




<t>PRF_HMAC_MD5 has been downgraded from MAY in RFC4307 to MUST NOT. There is an industry-wide trend to 
        deprecate its usage. MD5 is known to be subject to collision which can lead to transcript attacks.  

<c>3</c><c>DSS Digital Signature</c><c>SHOULD NOT</c><c></c>
<t>DSS Digital Signature is bound to SHA-1 and has the same level of security as 1024-bit RSA. It is expected to be downgraded to MUST NOT in the future.</t>                 

d) rsa

<t>RSA Digital Signature is mostly kept for interoperability. It is expected to be downgraded 
in the future as signatures are based on RSASSA-PKCS1-v1.5, not any more recommended. 
Instead, more robust use of RSA is expected to be performed via the Digital Signature method. 
RSA Digital Signature are not recommended to use keys lower that 2048. These key lengths have
not been further downgraded to remain coherent with the 1024-bit MODP Group whose status has been
set to SHOULD NOT. Unlike 1024-bit MODP, attacks on authentication requires an online attack and
attacks performed after the authentication phase have very little value. On the other hand,
attacks performed on the Diffie Hellman exchange after the initial exchange can lead to a 
complete decryption of the communication.</t>

e) psk

Do we have any recommendation for the authentication based on PSK?

f) oid

What are the OID values for RSA / ECDSA?

<c>OID</c><c>RSA</c><c>MUST</c><c>With keys length 2048</c>
<c>OID</c><c>RSA</c><c>SHOULD</c><c>With keys length 3072/4096</c>
<c>OID</c><c>RSA</c><c>SHOULD NOT</c><c>With keys length lower than 2048</c>
<c>OID</c><c>ECDSA</c><c>SHOULD</c><c></c> 

COMMENT: Shouldn't we rather remove MD5 from the Digital Signature Recommendation section ?

COMMENT: Currently SHA1 status is set to MUST. Shouldn't we set it to MUST- ?